requirements.txt
opencv-python==4.8.1.78

pyserial==3.5

numpy==1.24.3

pillow==10.0.1

psutil==5.9.5

pytest==7.4.2